Casting result to int

+++ math/0x02-calculus/9-sum_total.py
@@ -7,5 +7,5 @@ def summation_i_squared(n):
     """a function that calculates \sum_{i=1}^{n} i^2"""
     if  n < 1 or type(n) != int:
         return None
-    result = (n * (n + 1) * (2 * n + 1)) // 6
+    result = int((n * (n + 1) * (2 * n + 1)) // 6)
     return result
